jobs: change p_job variable type from short to int
This change addresses an issue where ksh hangs when the number of jobs exceeds `2^15 - 1` (32767), which is the maximum value for a `short`.

+++ src/cmd/ksh93/include/jobs.h
@@ -51,7 +51,7 @@ struct process
 	pid_t		p_pid;		/* process ID */
 	pid_t		p_pgrp;		/* process group */
 	pid_t		p_fgrp;		/* process group when stopped */
-	short		p_job;		/* job number of process */
+	int		p_job;		/* job number of process */
 	unsigned short	p_exit;		/* exit value or signal number */
 	unsigned short	p_exitmin;	/* minimum exit value for xargs */
 	unsigned short	p_flag;		/* flags - see below */
